requirements.txt
flask
torch
transformers
sentence-transformers
ultralytics
Pillow
opencv-python
absl-py==2.2.2
alembic==1.16.2
attrs==25.3.0
bidict==0.23.1
blinker==1.9.0
boto3==1.37.37
botocore==1.37.37
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.2
cycler==0.12.1
faiss-cpu==1.11.0
filelock==3.18.0
filetype==1.2.0
fire==0.7.0
Flask==3.1.0
Flask-Migrate==4.1.0
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.1.1
flatbuffers==25.2.10
fonttools==4.57.0
fsspec==2025.3.2
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.2.3
grpcio==1.71.0
h11==0.14.0
hf-xet==1.1.5
huggingface-hub==0.33.1
idna==3.7
itsdangerous==2.2.0
jax==0.6.0
jaxlib==0.6.0
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.1
kiwisolver==1.4.8
Mako==1.3.10
Markdown==3.8
MarkupSafe==3.0.2
matplotlib==3.10.1
mediapipe==0.10.21
ml_dtypes==0.5.1
mpmath==1.3.0
networkx==3.4.2
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
opencv-contrib-python==4.11.0.86
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
opt_einsum==3.4.0
packaging==25.0
pandas==2.2.3
pillow==11.2.1
pillow_heif==0.22.0
protobuf==4.25.6
psutil==7.0.0
py-cpuinfo==9.0.0
pybboxes==0.1.6
pycparser==2.22
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-engineio==4.12.0
python-socketio==5.13.0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
roboflow==1.1.61
s3transfer==0.11.5
safetensors==0.5.3
sahi==0.11.22
scikit-learn==1.7.0
scipy==1.15.2
seaborn==0.13.2
sentence-transformers==2.5.1
sentencepiece==0.2.0
shapely==2.1.0
simple-websocket==1.1.0
six==1.17.0
smmap==5.0.2
sounddevice==0.5.1
SQLAlchemy==2.0.41
sympy==1.13.1
tensorboard==2.19.0
tensorboard-data-server==0.7.2
termcolor==3.0.1
terminaltables==3.1.10
thop==0.1.1.post2209072238
threadpoolctl==3.6.0
tokenizers==0.15.2
tomli==2.2.1
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
transformers==4.39.3
triton==3.2.0
typing_extensions==4.13.2
tzdata==2025.2
ultralytics==8.3.111
ultralytics-thop==2.0.14
urllib3==2.4.0
Werkzeug==3.1.3
wsproto==1.2.0
